+++ test/diff_check.jl
@@ -4,8 +4,10 @@
 # in the LICENSE.md file or at https://opensource.org/licenses/MIT.
 
 using Test
+import SolverCore
+import LowRankOpt as LRO
 
-function test_vecprod(f, len, J; tol = 1e-6)
+function _test_vecprod(f, len, J; tol = 1e-6)
     v = ones(len)
     @test f(v) ≈ J * v rtol = tol atol = tol
     v = -ones(len)
@@ -33,7 +35,7 @@ function jac_check(model, x; kws...)
     f(x) = NLPModels.cons(model, x)
     J = FiniteDiff.finite_difference_jacobian(f, x)
     @testset "jprod" begin
-        test_vecprod(
+        _test_vecprod(
             v -> NLPModels.jprod(model, x, v),
             model.meta.nvar,
             J;
@@ -41,7 +43,7 @@ function jac_check(model, x; kws...)
         )
     end
     @testset "jtprod" begin
-        test_vecprod(
+        _test_vecprod(
             v -> NLPModels.jtprod(model, x, v),
             model.meta.ncon,
             J';
@@ -56,7 +58,7 @@ function hess_check(model, x; kws...)
     f(x) =
         obj_weight * NLPModels.obj(model, x) - dot(y, NLPModels.cons(model, x))
     J = FiniteDiff.finite_difference_hessian(f, x)
-    return test_vecprod(
+    return _test_vecprod(
         v -> NLPModels.hprod(model, x, y, v; obj_weight),
         model.meta.nvar,
         J;
@@ -88,4 +90,67 @@ function diff_check(model)
     @testset "Hessian" begin
         hess_check(bm, x)
     end
+end
+
+struct ConvexSolver{T} <: SolverCore.AbstractOptimizationSolver
+    model::LRO.Model{T}
+    stats::SolverCore.GenericExecutionStats{T,Vector{T},Vector{T},Any}
+end
+
+function ConvexSolver(model::LRO.Model)
+    stats = SolverCore.GenericExecutionStats(model)
+    return ConvexSolver(model, stats)
+end
+
+function LRO.MOI.get(solver::ConvexSolver, ::LRO.Solution)
+    return LRO.VectorizedSolution(solver.stats.solution, solver.model.dim)
+end
+
+function SolverCore.solve!(::ConvexSolver, ::LRO.Model)
+    return
+end
+
+function _alloc_schur_complement(model, i, Wi, H)
+    if VERSION < v"1.11"
+        return
+    end
+    LRO.add_schur_complement!(model, i, Wi, H)
+    @test 0 == @allocated LRO.add_schur_complement!(model, i, Wi, H)
+end
+
+function schur_test(model::LRO.BufferedModelForSchur{T}, w) where {T}
+    n = model.meta.ncon
+    y = rand(T, n)
+
+    Jv = similar(y)
+    vJ = similar(w)
+    NLPModels.jprod!(model, w, w, Jv)
+    NLPModels.jtprod!(model, w, y, vJ)
+    @test dot(Jv, y) ≈ dot(vJ, w)
+
+    H = zeros(n, n)
+    H = LRO.schur_complement!(model, w, H)
+    Hy = similar(y)
+    LRO.eval_schur_complement!(model, w, y, Hy)
+    @test Hy ≈ H * y
+    for i in LRO.matrix_indices(model)
+        Wi = @inferred w[i]
+        _alloc_schur_complement(model, i, Wi, H)
+    end
+    for i in LRO.matrix_indices(model)
+        ret = LRO.dual_cons!(model, y, i)
+        @test ret isa SparseMatrixCSC
+    end
+    dcons = ones(LRO.num_scalars(model))
+    LRO.dual_cons!(model, y, dcons, LRO.ScalarIndex)
+    @test dcons ≈ model.model.d_lin - model.model.C_lin' * y
+end
+
+function schur_test(model::LRO.Model{T}, κ) where {T}
+    w = rand(T, model.meta.nvar)
+    W = LRO.VectorizedSolution(w, model.dim)
+    for i in LRO.matrix_indices(model)
+        W[i] .= W[i] .+ W[i]'
+    end
+    return schur_test(LRO.BufferedModelForSchur(model, κ), W)
 end
